currency_converter
This is an Android app I developed for currency exchange and weather forecast. I developed it as I was about to go to travel and I think using this app could help me out a lot throughout the travel by converting currency easily and getting to know the upcoming weather. I used some free APIs on the internet(they may not work as of now since they may stop the free services anytime) and Java skills to develop this app. Despite the UI is not attractive, it is useful and comes in handy while travelling.